Moved all cronjobs and automated email to email.js
